test: split history tests into multiple files
